telegram_bot/requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
black==24.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
fastapi==0.110.0
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
magic-filter==1.0.12
Mako==1.3.2
MarkupSafe==2.1.5
multidict==6.0.5
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
pydantic==2.5.3
pydantic-settings==2.2.1
pydantic_core==2.14.6
pypng==0.20220715.0
PyQRCode==1.2.1
python-dotenv==1.0.1
python-telegram-bot==21.0.1
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.36.3
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.29.0
yarl==1.9.4